Inline ResponseBodyCommon
This is an artifact of our multiplatform exploration.
